apps/load-generator/requirements.txt
requests
pyyaml
pyopenssl
prometheus_client
datetime
